# map: avoid allocations during batch lookup of common types