Extend the optimization from 4609dc73 ("map: zero-allocation operations for common types") to BatchLookup. This means that types without implicit padding will not incur allocations.

    core: 1
    goos: linux
    goarch: amd64
    pkg: github.com/cilium/ebpf
    cpu: 12th Gen Intel(R) Core(TM) i7-1260P
                        │   base.txt   │              opt.txt               │
                        │    sec/op    │   sec/op     vs base               │
    Iterate/BatchLookup   4317.0n ± 2%   771.6n ± 1%  -82.13% (p=0.002 n=6)

                        │   base.txt    │              opt.txt              │
                        │     B/op      │    B/op     vs base               │
    Iterate/BatchLookup   16432.00 ± 0%   48.00 ± 0%  -99.71% (p=0.002 n=6)

                        │  base.txt  │              opt.txt              │
                        │ allocs/op  │ allocs/op   vs base               │
    Iterate/BatchLookup   4.000 ± 0%   2.000 ± 0%  -50.00% (p=0.002 n=6)