@ti-mo 2023-12-05
Nice. Can we also enforce this via `testing.AllocsPerRun()`? (TIL! Was looking for something like this for the longest time.)